Fixed AddOn bug. Added new class PetHelper to be called by scripts to help resolve some issues. Added power to auras to better calculate damage.

+++ WoWPetBattler/PetTeam.cpp
@@ -76,7 +76,7 @@ quint8 PetTeam::GetActivePetIndex()
 }
 
 //Get the number of pets on the team.
-int PetTeam::GetNumPets()
+quint8 PetTeam::GetNumPets()
 {
 	return this->pets.size()-1;
 }
